Renamed RepositoryRepo to Repo and removed old Repo interface
The old hierarchy resulted in confusing RepositoryRepo name
