+++ src/org/rascalmpl/util/maven/MirrorRepo.java
@@ -3,7 +3,7 @@ package org.rascalmpl.util.maven;
 import org.apache.maven.model.Repository;
 import org.apache.maven.settings.Mirror;
 
-public class MirrorRepo extends RepositoryRepo {
+public class MirrorRepo extends Repo {
     private Mirror mirror;
 
     public MirrorRepo(Mirror mirror, Repository repository) {
