+++ src/org/rascalmpl/util/maven/Repo.java
@@ -2,12 +2,26 @@ package org.rascalmpl.util.maven;
 
 import org.apache.maven.model.Repository;
 
-public interface Repo {
-    Repository getRepository();
+public class Repo {
+    private Repository repository;
 
-    String getId();
+    public Repo(Repository repository) {
+        this.repository = repository;
+    }
 
-    String getUrl();
-    
-    String getLayout();
+    public Repository getMavenRepository() {
+        return repository;
+    }
+
+    public String getId() {
+        return repository.getId();
+    }
+
+    public String getLayout() {
+        return repository.getLayout();
+    }
+
+    public String getUrl() {
+        return repository.getUrl();
+    }
 }
